feat(core): governance models, authorisation predicates and privacy gate suite
Zone 2 Django service: five-store models behind a router, an append-only
hash-chained audit log, the consent ledger, and the RBAC/ABAC access predicates.

Two guarantees are enforced structurally rather than asserted:

Tamper evidence. audit_event and consent_entry carry BEFORE UPDATE OR DELETE
triggers, risk_assessment permits only the superseded_by back-link to change,
and manobal_audit_chain_break() verifies the whole chain in one pass. The tests
disable the trigger and excise a row to prove the chain names the break point,
and one passing test documents the limit honestly: a full rewrite verifies,
which is why AuditAnchor exists.

No score escapes. RiskAssessmentRecord has no numeric column and the gate suite
introspects the model to keep it that way, so adding one fails the build. The
redaction filter treats 'wsi' and 'domain_scores' as sensitive because a log
line is a way out of the engine too.

Access is an event with an expiry, not a property of a role. Every
individual-level read needs role, force, unit subtree, an officer-visible tier,
a live grant and a real second factor; AccessGrant.save clamps any expiry past
the fourteen-day ceiling rather than trusting callers.

Notable during the build:

- The privacy gate initially flagged Unit.name and was wrong to. A column is
  dangerous because of the row it sits on, so the gate now scopes to tables
  carrying subject_token, with a second test asserting that set is non-empty so
  it cannot pass vacuously.
- self.pk is not None is the wrong insert test; _state.adding is Django's own
  and stays correct when a key is assigned before first save.
- OIDC settings are parsed once into a frozen typed config that rejects
  alg=none, moving misconfiguration to process start rather than to an
  authentication attempt.
- bio_store and voice_store apps exist with grants and no tables pending
  TimescaleDB and pgvector. Coverage renormalisation already handles absent
  domains, so nothing downstream is blocked.

ruff and mypy --strict clean across 90 source files; 323 tests pass.

Refs: SDD v1.0 sections 3.2, 4.5, 5.1, 5.2, 5.4, 6.2, 6.3, 7.3, 7.9, 9.6, 10.1

+++ packages/manobal-synth/src/manobal_synth/cohorts.py
@@ -28,6 +28,7 @@ from collections.abc import Sequence
 from dataclasses import dataclass
 
 import numpy as np
+
 from manobal_risk import AcuteTriggerKind
 
 from .arrays import FloatArray
@@ -154,7 +155,7 @@ def _draw_gaming(
 ) -> np.ndarray:
     """Select suppressors, preferring subjects who are already deteriorating."""
     count = len(distressed)
-    target = int(round(config.gaming_cohort * count))
+    target = round(config.gaming_cohort * count)
     gaming = np.zeros(count, dtype=bool)
     if target <= 0:
         return gaming
@@ -178,11 +179,11 @@ def _draw_acute(
 ) -> tuple[dict[int, int], dict[int, AcuteTriggerKind]]:
     """Choose who has an acute event, when, and of which kind."""
     count = len(distressed)
-    target = int(round(config.acute_events * count))
+    target = round(config.acute_events * count)
     if target <= 0:
         return {}, {}
 
-    from_distressed = min(int(round(target * _ACUTE_DISTRESSED_SHARE)), int(distressed.sum()))
+    from_distressed = min(round(target * _ACUTE_DISTRESSED_SHARE), int(distressed.sum()))
     chosen = list(rng.permutation(np.flatnonzero(distressed))[:from_distressed])
     remaining = target - len(chosen)
     if remaining > 0:
